requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
channels==4.3.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.2
